# bam-api/src/main/java/com/account/bam/model/account/balance/Balance.java
package com.account.bam.model.account.balance;

import com.account.bam.model.account.Account;

import java.math.BigDecimal;
import java.time.LocalDateTime;
import javax.persistence.*;

//todo: define composite unique key relationship here as well
@Entity
public class Balance {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "Id")
    private Long id;

    @Column(name = "Amount", precision = 18, scale = 4, nullable = false)
    private BigDecimal amount;

    @Column(name = "Currency", length = 3, nullable = false)
    private String currency;

    @Column(name = "CreationTime")
    private LocalDateTime creationTime;

    @Column(name = "UpdateTime")
    private LocalDateTime updateTime;

    @ManyToOne(fetch=FetchType.EAGER)
    @JoinColumn(name="AccountId")
    private Account account;


    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public BigDecimal getAmount() {
        return amount;
    }

    public void setAmount(BigDecimal amount) {
        this.amount = amount;
    }

    public String getCurrency() {
        return currency;
    }

    public void setCurrency(String currency) {
        this.currency = currency;
    }

    public LocalDateTime getCreationTime() {
        return creationTime;
    }

    public void setCreationTime(LocalDateTime creationTime) {
        this.creationTime = creationTime;
    }

    public LocalDateTime getUpdateTime() {
        return updateTime;
    }

    public void setUpdateTime(LocalDateTime updateTime) {
        this.updateTime = updateTime;
    }

    public Account getAccount() {
        return account;
    }

    public void setAccount(Account account) {
        this.account = account;
    }
}